move `styled_button_toggleable` to notedeck_ui
